pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/lessrest/bubble"
"Bug Tracker" = "https://github.com/lessrest/bubble/issues"
Repository = "https://github.com/lessrest/bubble"

[project]
name = "bubble"
version = "0.2.0"
description = "a Node.Town bubble server"
authors = [{ name = "Mikael Brockman", email = "mikael@brockman.se" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = "==3.12.*"
license = { file = "LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: Trio",
    "Programming Language :: Python :: 3.12",
]
keywords = ["bubble", "rdf", "logic", "llm", "git", "async", "trio"]

dependencies = [
    # Structured concurrency library for managing async code
    "trio==0.27.*",
    # Python library for working with RDF (Resource Description Framework)
    "rdflib==7.*",
    # Data validation using type annotations
    "pydantic==2.10.*",
    # Modern, fast web framework for building APIs with Python
    "fastapi==0.115.*",
    # ASGI framework used by FastAPI
    "starlette==0.41.*",
    "python-multipart==0.0.*",
    # ASGI server for Trio
    "hypercorn==0.17.*",
    # Rich text and beautiful formatting in the terminal
    "rich==13.*",
    # CLI framework; the FastAPI of CLI apps
    "typer==0.*",
    # Modern async HTTP client
    "httpx==0.*",
    # Various AI inference SDKs
    "replicate==1.*",
    "openai==1.55.*",
    "anthropic==0.39.*",
    # Better date and time manipulation for Python
    "arrow==1.*",
    # Parse and pretty-print HTML
    "beautifulsoup4==4.12.*",
    # Globally unique ID generator
    "py-xid==1.*",
    # Cross-platform unique machine ID generator
    "py-machineid==0.*",
    # Cross-platform library for retrieving system information
    "psutil==6.*",
    "structlog>=24.4.0",
    "pyogg>=0.6.14a1",
    "pydub>=0.25.1",
    "deepgram-sdk>=3.7.7",
    "trio-websocket>=0.11.1",
    "backend>=0.2.4.1",
    "ipykernel>=6.29.5",
    "pip>=24.3.1",
    "cbor2>=5.6.5",
    "swash>=0.1.0",
    "invoke>=2.2.0",
    "pytest>=8.3.3",
    "pytest-httpx>=0.34.0",
    "hypothesis>=6.119.4",
    "pytest-xdist>=3.6.1",
    "pytest-trio>=0.8.0",
    "colorama>=0.4.6",
    "cryptography>=44.0.0",
    "asgi-lifespan>=2.1.0",
    "pyld>=2.0.4",
    "importhook>=1.0.9",
    "tenacity>=9.0.0",
    "pillow>=11.0.0",
    "google-genai>=0.2.1",
    "wn>=0.11.0",
    "pudb>=2024.1.3",
    "trio-asyncio>=0.15.0",
    "httpx-ws>=0.7.0",
]

[tool.uv.sources]
# stay current with rdflib
rdflib = { git = "https://github.com/RDFLib/rdflib" }
swash = { workspace = true }

[tool.uv.workspace]
members = ["swash"]

[project.scripts]
bubble = "bubble.cli:app"

[tool.uv]
package = true
dev-dependencies = [
    # Linter and formatter
    "ruff==0.*",
    # Static type checker
    "pyright==1.*",
    # Task runner, see tasks.py
    "invoke==2.2.*",
    # Property-based testing framework
    "hypothesis==6.*",
    # Mocking and testing library
    "mock==5.*",
    # Popular Python testing framework
    "pytest==8.*",
    # Pytest plugin for testing async code using Trio
    "pytest-trio==0.*",
    # Generates Markdown reports from pytest results
    "pytest-md-report==0.6.*",
    # Auto-rerun pytest on file changes
    "pytest-watcher==0.*",
    # Thin wrapper around the mock package for pytest
    "pytest-mock==3.*",
    # Pytest plugin for mocking httpx requests
    "pytest-httpx==0.*",
    # Run tests in parallel
    "pytest-xdist==3.6.1",
    # Coverage reporting plugin for pytest
    "pytest-cov==6.*",
    # Code coverage measurement tool
    "coverage==7.*",
    # Used only for comment wrapping functionality
    "autopep8==2.*",
]

[tool.ruff]
line-length = 76
target-version = "py312"

[tool.ruff.lint.isort]
lines-between-types = 1
length-sort = true

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
trio_mode = true
md_report_verbose = 1
mock_use_standalone_module = true
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

# Run tests in parallel
# addopts = "-n 4"

[tool.hatch.metadata]
allow-direct-references = true
